Replace removed config options from 5.5 and old 5.6 installations
The two options key-buffer and myisam-recover were replaced in 5.5
with key-buffer-size and myisam-recover-options. In 5.7 the unique-prefix
option that allowed these to work in 5.6 was also removed. If the user
has customized the old config file it will not be replaced and the server
will fail to start.
Renaming these two options in /etc/mysql/my.cnf should clear up a common
upgrade error.
If the config file is changed, we create a backup in
/etc/mysql/my.cnf.migrated.bak (LP: #1571865)

Reviewed-by: Martin Pitt
Edited-by: Robie Basak (match space at start of line as suggested by
 Martin; tweak version match string for current version status)