Set perl to include working dir for d/tests/upstream tests
As part the fix for https://www.debian.org/security/2016/dsa-3628 perl will
no longer include the current working directory on the include search path.
The mtr suite will fail because it includes paths relative to the current dir.
A workaround until a more complete fix of the mtr script is to set perl to
include the current dir (mtr suite is run from a dir that's only root-writeable)