fix_old_config_options: adjust version comparison
In:

> if dpkg --compare-versions "$2" le-nl "5.7.13-0ubuntu3~"

"$2" may be empty if mysql-server-5.7 has never been installed before,
for example when doing a release upgrade from a previous release that
shipped MySQL 5.5 or 5.6. In this case, fix_old_config_options will
never run, when we do actually want it to do so.

By changing this to le, an empty "$2" will compare lower, and so
fix_old_config_options will run in this case.

Since fix_old_config_options is idempotent, the extra run in case of the
user not having a mysql.cnf previously will be a no-op. So this solution
is better than missing this particular edge case.