Add patch for failing test datadir_permission
The test is failing during the build (not dep8) because it uses
a custom config which tries to load a mysqlx related env variable
which might not be available. The test does not appear to need this
custom config, so we patch it out until the issue can be fixed upstream.