Added functions for starting and stopping the server without invoke
We want to stop using invoke-rc.d to start and stop the server during
init, since it does not work if the user has disabled the service.
Added functions for starting and stopping the server with and without
authentication disabled (--skip-grant-tables), and for stopping it
using a kill signal to the pid.
The server is started with the --daemonize flag, which makes the
process fork once the server is ready for use.