Add maintainer script freeze mode
See doc/freeze-mode/README in src:mysql-defaults 1.0.4 for rationale.

Bump the mysql-server-5.7 depends on mysql-common so the documentation
will be available when we symlink /etc/mysql/FROZEN. The presence of the
symlink is sufficient for all maintainer scripts; the target is only
needed to be present for the user for documentation purposes. So this
only needs to be a Depends and not a Pre-Depends.