Add NEWS for 5.7.20-2
(cherry picked from commit 438c40229bdaf0a3784f02fdb01a3d142b0714e9)

This should have been incorporated in Lars' merge earlier, but he had
started the merge before I thought it'd be a good idea to add to the
Debian upload. So I'm cherry-picking it here to fix up the Ubuntu merge
so that the "deletion" of the NEWS file isn't part of the Ubuntu delta.
It doesn't seem worth redoing the merge just for this.