Adjust changelog
It feels misleading to me when bug is closed as part of the announcement
of the package merge, since that makes it look like it was the merge bug
when it was not.

Separately, we should add a message noting that the previous delta
introduced by Ubuntu has been dropped. In this case, we've dropped the
dropping of the patch from the delta. That we've also fixed the
underlying issue is established in the part of the changelog from the
Debian uploads.