d/lintian-overrides: Add json override for router license
Similar to the existing one for server, router contains no rapidjson
code with the json license, but refers to the full rapidjson license.

Everything except the header files has been removed from the extra/rapidjson
directory, and the file extra/rapidjson/license_readme-oracle.txt contains
and explanation of the issue and resolution.

From the extra/rapidjson/license.txt file:
"
To avoid the problematic JSON license in your own projects,
it's sufficient to exclude the bin/jsonchecker/ directory,
as it's the only code under the JSON license
"