d/patches: Fix segmentation fault when libmysqlclient mistakenly reads mariadb charsets
This happens because mariadb installs charset files to the same place MySQL does, so if
mariadb is combined with libmysqlclient (and possibly vice-versa), the client library
may attempt to use those charset files.

A bug in the parser combined with the differences in charset data can lead to the client
library segfaulting. The patch simply prevents the segfault by checking that the collation
pointer is not NULL, until the underlying issue in the parser is fixed upstream.

Ref: https://bugs.launchpad.net/ubuntu/+source/mysql-8.0/+bug/1877504